requirements.txt
# /requirements.txt

connexion[swagger-ui]
flask
python-dotenv
python-jose
six